+++ packages/tasks/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @workspace/tasks
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @workspace/database@0.1.3
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @workspace/database@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes